completely reqorked the tone rules
The tone rules were seriously broken due to my incomplete understanding
complicated by the fact that each book seems to have a slightly
different idea of what they are supposed to be. Using the concept of
dead and live syllables, this should come pretty close to reality now.
It still needs cleaning up of the tests and the IPA romanization that
relies on certain tone names.